+++ CHANGELOG.md
@@ -1,8 +1,10 @@
+* adjusting tests to the new API
+* handling differences between node/broser worker
+* cleaning up repo
+
+
 <a name="0.2.12"></a>
 ## [0.2.12](https://github.com/hbi99/QureJS.js/compare/0.2.11...v0.2.12) (2015-11-15)
-
-
-
 
 * qure.js node:ifed