+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* demo files for browser environment adjusted
 * handling non-node environment kill signal
 * adjusting tests to the new API
 * handling differences between node/broser worker