+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* precede pipe from node worker
 * added pause resume to node thread worker
 * renaming from "WRK_*" to "*Worker"